move load scalar to data.py file
